refactor(cli): move cli & interacitve to seperate files
